# api/openapi.yaml
openapi: 3.1.0
info:
    title: 'Voting API'
    version: 1.0.0
    description: 'Secure voting system API — Express, TypeScript, Prisma. Auto-generated from JSDoc comments.'
servers:
    -
        url: 'http://localhost:4000'
        description: 'Local development'
    -
        url: 'https://api.yourdomain.com'
        description: Production
components:
    securitySchemes:
        bearerAuth:
            type: http
            scheme: bearer
            bearerFormat: JWT
            description: 'Paste an **access token** here.'
security:
    -
        bearerAuth: []
...
paths:
...
    /api/voters:
        get:
            tags:
                - Voters
            summary: 'Get all voters'
            description: 'Retrieve a paginated list of voters with optional filtering and search'
            security:
                -
                    bearerAuth: []
            parameters:
                -
                    name: page
                    in: query
                    description: 'Page number for pagination'
                    schema:
                        type: string
                        default: '1'
                -
                    name: limit
                    in: query
                    description: 'Number of voters per page'
                    schema:
                        type: string
                        default: '50'
                -
                    name: search
                    in: query
                    description: 'Search term for name, phone, or unique identifier'
                    schema:
                        type: string
                -
                    name: status
                    in: query
                    description: 'Filter by voter status'
                    schema:
                        type: string
                -
                    name: classYearGroup
                    in: query
                    description: 'Filter by class year group'
                    schema:
                        type: string
            responses:
                '200':
                    description: 'Voters retrieved successfully'
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    voters: {type: array, items: {type: object, properties: {id: {type: string, format: uuid}, fullName: {type: string}, phone: {type: string}, uniqueIdentifier: {type: string}, status: {type: string}, classYearGroup: {type: string}, isVerified: {type: boolean}, createdAt: {type: string, format: date-time}}}}
                                    pagination: {type: object, properties: {page: {type: number}, limit: {type: number}, total: {type: number}, totalPages: {type: number}}}
                '401':
                    description: 'Unauthorized - Authentication required'
                '403':
                    description: 'Forbidden - EC access required'
        post:
            tags:
                - Voters
            summary: 'Create new voter'
            description: 'Create a new voter with the provided information'
            security:
                -
                    bearerAuth: []
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            type: object
                            required:
                                - fullName
                                - phone
                                - classYearGroup
                                - uniqueIdentifier
                            properties:
                                fullName:
                                    type: string
                                    minLength: 2
                                    description: 'Full name of the voter'
                                phone:
                                    type: string
                                    description: 'Ghana phone number'
                                classYearGroup:
                                    type: string
                                    description: 'Class year group of the voter'
                                uniqueIdentifier:
                                    type: string
                                    description: 'Unique identifier for the voter'
                                status:
                                    type: string
                                    description: 'Status of the voter'
                                    default: active
            responses:
                '201':
                    description: 'Voter created successfully'
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    id: {type: string, format: uuid}
                                    fullName: {type: string}
                                    phone: {type: string}
                                    classYearGroup: {type: string}
                                    uniqueIdentifier: {type: string}
                                    status: {type: string}
                                    createdAt: {type: string, format: date-time}
                '400':
                    description: 'Validation error'
                '401':
                    description: 'Unauthorized - Authentication required'
                '403':
                    description: 'Forbidden - EC access required'
                '409':
                    description: 'Conflict - Phone number already exists'
    '/api/voters/{id}':
        put:
            tags:
                - Voters
            summary: 'Update voter'
            description: 'Update an existing voter''s information'
            security:
                -
                    bearerAuth: []
            parameters:
                -
                    name: id
                    in: path
                    required: true
                    description: 'Voter ID'
                    schema:
                        type: string
                        format: uuid
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            type: object
                            required:
                                - fullName
                                - phone
                                - classYearGroup
                                - uniqueIdentifier
                            properties:
                                fullName:
                                    type: string
                                    minLength: 2
                                    description: 'Full name of the voter'
                                phone:
                                    type: string
                                    description: 'Ghana phone number'
                                classYearGroup:
                                    type: string
                                    description: 'Class year group of the voter'
                                uniqueIdentifier:
                                    type: string
                                    description: 'Unique identifier for the voter'
                                status:
                                    type: string
                                    description: 'Status of the voter'
            responses:
                '200':
                    description: 'Voter updated successfully'
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    id: {type: string, format: uuid}
                                    fullName: {type: string}
                                    phone: {type: string}
                                    classYearGroup: {type: string}
                                    uniqueIdentifier: {type: string}
                                    status: {type: string}
                                    updatedAt: {type: string, format: date-time}
                '400':
                    description: 'Validation error'
                '401':
                    description: 'Unauthorized - Authentication required'
                '403':
                    description: 'Forbidden - EC access required'
                '404':
                    description: 'Voter not found'
                '409':
                    description: 'Conflict - Phone number already exists'
        delete:
            tags:
                - Voters
            summary: 'Delete voter'
            description: 'Soft delete a voter (marks as inactive)'
            security:
                -
                    bearerAuth: []
            parameters:
                -
                    name: id
                    in: path
                    required: true
                    description: 'Voter ID'
                    schema:
                        type: string
                        format: uuid
            responses:
                '200':
                    description: 'Voter deleted successfully'
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    message: {type: string, example: 'Voter deleted successfully'}
                '401':
                    description: 'Unauthorized - Authentication required'
                '403':
                    description: 'Forbidden - EC access required'
                '404':
                    description: 'Voter not found'
...